+++ client/src/city_ui.rs
@@ -1,31 +1,28 @@
 use macroquad::color::BLACK;
-use std::cmp;
-use std::collections::HashMap;
-use std::f32::consts::PI;
-
 use macroquad::hash;
-use macroquad::math::{f32, i32, u32, vec2};
+use macroquad::math::{i32, u32, vec2};
 use macroquad::prelude::*;
-use macroquad::ui::{root_ui, Ui};
+use macroquad::ui::root_ui;
 use server::city::{City, MoodState};
 use server::city_pieces::Building;
-use server::game::{Action, Game};
+use server::game::Game;
 use server::player::Player;
-use server::playing_actions::PlayingAction;
 use server::position::Position;
-use server::resource_pile::{PaymentOptions, ResourcePile};
+use server::resource_pile::PaymentOptions;
+use std::cmp;
+use std::collections::HashMap;
 
+use crate::construct_ui::add_construct_button;
+use crate::happiness_ui::increase_happiness_click;
 use crate::hex_ui::pixel_to_coordinate;
-use crate::payment_ui::{
-    new_resource_map, payment_dialog, HasPayment, Payment, ResourcePayment, ResourceType,
-};
-use crate::ui::{can_play_action, IncreaseHappiness, Point, State};
-use crate::{hex_ui, ui, ActiveDialog};
+use crate::payment_ui::{new_resource_map, HasPayment, Payment, ResourcePayment, ResourceType};
+use crate::ui::{can_play_action, State};
+use crate::{hex_ui, influence_ui, ui, ActiveDialog};
 
 pub struct CityMenu<'a> {
-    player_index: usize,
-    city_owner_index: usize,
-    city_position: &'a Position,
+    pub player_index: usize,
+    pub city_owner_index: usize,
+    pub city_position: &'a Position,
 }
 
 impl<'a> CityMenu<'a> {
@@ -57,15 +54,15 @@ impl<'a> CityMenu<'a> {
 }
 
 pub struct ConstructionPayment {
-    player_index: usize,
-    city_position: Position,
-    city_piece: Building,
-    payment: Payment,
-    payment_options: PaymentOptions,
+    pub player_index: usize,
+    pub city_position: Position,
+    pub city_piece: Building,
+    pub payment: Payment,
+    pub payment_options: PaymentOptions,
 }
 
 impl ConstructionPayment {
-    fn new(
+    pub fn new(
         game: &Game,
         player_index: usize,
         city_position: Position,
@@ -122,15 +119,7 @@ pub fn show_city_menu(game: &mut Game, menu: CityMenu) -> Option<ActiveDialog> {
 
     root_ui().window(hash!(), vec2(30., 700.), vec2(500., 200.), |ui| {
         ui.label(None, &menu.city_position.to_string());
-
-        let closet_city_pos = &menu
-            .get_player(game)
-            .cities
-            .iter()
-            .min_by_key(|c| c.position.distance(menu.city_position))
-            .unwrap()
-            .position
-            .clone();
+        let closest_city_pos = &influence_ui::closest_city(&game, &menu);
 
         for (building, name) in building_names() {
             if can_play_action(game) {
@@ -138,114 +127,18 @@ pub fn show_city_menu(game: &mut Game, menu: CityMenu) -> Option<ActiveDialog> {
                     let _ = result.insert(d);
                 }
 
-                add_influence_button(game, &menu, ui, closet_city_pos, &building, name);
+                influence_ui::add_influence_button(
+                    game,
+                    &menu,
+                    ui,
+                    closest_city_pos,
+                    &building,
+                    name,
+                );
             };
         }
     });
     result
-}
-
-fn add_construct_button(
-    game: &Game,
-    menu: &CityMenu,
-    ui: &mut Ui,
-    building: &Building,
-    name: &str,
-) -> Option<ActiveDialog> {
-    let owner = menu.get_city_owner(game);
-    let city = menu.get_city(game);
-    if (menu.is_city_owner())
-        && city.can_construct(building, owner)
-        && ui.button(None, format!("Build {}", name))
-    {
-        return Some(ActiveDialog::ConstructionPayment(ConstructionPayment::new(
-            game,
-            menu.player_index,
-            menu.city_position.clone(),
-            building.clone(),
-        )));
-    }
-    None
-}
-
-fn add_influence_button(
-    game: &mut Game,
-    menu: &CityMenu,
-    ui: &mut Ui,
-    closet_city_pos: &Position,
-    building: &Building,
-    building_name: &str,
-) {
-    if !menu.get_city(game).city_pieces.can_add_building(building) {
-        let start_position = if menu.is_city_owner() {
-            menu.city_position
-        } else {
-            closet_city_pos
-        };
-        if let Some(cost) = game.influence_culture_boost_cost(
-            menu.player_index,
-            start_position,
-            menu.city_owner_index,
-            menu.city_position,
-            building,
-        ) {
-            if ui.button(
-                None,
-                format!("Attempt Influence {} for {}", building_name, cost),
-            ) {
-                game.execute_action(
-                    Action::PlayingAction(PlayingAction::InfluenceCultureAttempt {
-                        starting_city_position: start_position.clone(),
-                        target_player_index: menu.city_owner_index,
-                        target_city_position: menu.city_position.clone(),
-                        city_piece: building.clone(),
-                    }),
-                    menu.player_index,
-                );
-            }
-        }
-    }
-}
-
-pub fn pay_construction_dialog(game: &mut Game, payment: &mut ConstructionPayment) -> bool {
-    payment_dialog(
-        payment,
-        |cp| cp.payment.get(ResourceType::Discount).current == 0,
-        |cp| {
-            game.execute_action(
-                Action::PlayingAction(PlayingAction::Construct {
-                    city_position: cp.city_position.clone(),
-                    city_piece: cp.city_piece.clone(),
-                    payment: cp.payment.to_resource_pile(),
-                    temple_bonus: None,
-                }),
-                cp.player_index,
-            )
-        },
-        |ap, r| match r {
-            ResourceType::Gold => ap.payment_options.gold_left > 0,
-            ResourceType::Discount => ap.payment_options.discount > 0,
-            _ => ap.payment.get(r).max > 0,
-        },
-        |cp, r| {
-            let gold = cp.payment.get_mut(ResourceType::Gold);
-            if gold.current > 0 {
-                gold.current -= 1;
-            } else {
-                cp.payment.get_mut(ResourceType::Discount).current += 1;
-            }
-            cp.payment.get_mut(r).current += 1;
-        },
-        |cp, r| {
-            let discount = cp.payment.get_mut(ResourceType::Discount);
-            if discount.current > 0 {
-                discount.current -= 1;
-            } else {
-                cp.payment.get_mut(ResourceType::Gold).current += 1;
-            }
-            cp.payment.get_mut(r).current -= 1;
-        },
-    )
 }
 
 pub fn draw_city(owner: &Player, city: &City, state: &State) {
@@ -277,7 +170,7 @@ pub fn draw_city(owner: &Player, city: &City, state: &State) {
     let mut i = 0;
     for player_index in 0..4 {
         for b in city.city_pieces.buildings(Some(player_index)).iter() {
-            let p = rotate_around(c, 30.0, 90 * i);
+            let p = hex_ui::rotate_around(c, 30.0, 90 * i);
             draw_text(
                 building_symbol(b),
                 p.x - 12.0,
@@ -287,14 +180,6 @@ pub fn draw_city(owner: &Player, city: &City, state: &State) {
             );
             i += 1;
         }
-    }
-}
-
-fn rotate_around(center: Point, radius: f32, angle_deg: i32) -> Point {
-    let angle_rad = PI / 180.0 * (angle_deg as f32);
-    Point {
-        x: center.x + radius * f32::cos(angle_rad),
-        y: center.y + radius * f32::sin(angle_rad),
     }
 }
 
@@ -352,68 +237,4 @@ fn city_click(state: &mut State, player: &Player, city: &City) {
         state.clear();
         state.focused_city = Some((player.index, pos.clone()));
     }
-}
-
-fn increase_happiness_click(
-    player: &Player,
-    city: &City,
-    pos: &Position,
-    increase_happiness: &IncreaseHappiness,
-) -> IncreaseHappiness {
-    let mut total_cost = increase_happiness.cost.clone();
-    let new_steps = increase_happiness
-        .steps
-        .iter()
-        .map(|(p, steps)| {
-            let old_steps = *steps;
-            if p == pos {
-                if let Some(r) = increase_happiness_steps(player, city, &total_cost, old_steps) {
-                    total_cost = r.1;
-                    return (p.clone(), r.0);
-                };
-            }
-            (p.clone(), old_steps)
-        })
-        .collect();
-
-    IncreaseHappiness::new(new_steps, total_cost)
-}
-
-fn increase_happiness_steps(
-    player: &Player,
-    city: &City,
-    total_cost: &ResourcePile,
-    old_steps: u32,
-) -> Option<(u32, ResourcePile)> {
-    if let Some(value) =
-        increase_happiness_new_steps(player, city, total_cost, old_steps, old_steps + 1)
-    {
-        return Some(value);
-    }
-    if let Some(value) = increase_happiness_new_steps(player, city, total_cost, old_steps, 0) {
-        return Some(value);
-    }
-    None
-}
-
-fn increase_happiness_new_steps(
-    player: &Player,
-    city: &City,
-    total_cost: &ResourcePile,
-    old_steps: u32,
-    new_steps: u32,
-) -> Option<(u32, ResourcePile)> {
-    if let Some(new_cost) = city.increase_happiness_cost(new_steps) {
-        let mut new_total = total_cost.clone();
-        if old_steps > 0 {
-            new_total -= city
-                .increase_happiness_cost(old_steps)
-                .expect("invalid steps");
-        }
-        new_total += new_cost;
-        if player.resources().can_afford(&new_total) {
-            return Some((new_steps, new_total));
-        }
-    }
-    None
 }
