+++ client/src/main.rs
@@ -4,8 +4,11 @@ use crate::ui::{ActiveDialog, State};
 
 mod advance_ui;
 mod city_ui;
+mod construct_ui;
 mod game_loop;
+mod happiness_ui;
 mod hex_ui;
+mod influence_ui;
 mod local_ui;
 mod log_ui;
 mod map_ui;
